refactor: remove fragments and xml navigation
