+++ presentation/src/main/java/com/example/alpacamusclemaintenance/presentation/MainActivity.kt
@@ -17,7 +17,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             MaterialTheme {
-                AlpacaMuscleMaintenanceApp()
+                App()
             }
         }
     }
