+++ presentation/src/main/java/com/example/alpacamusclemaintenance/presentation/feed/FeedView.kt
@@ -1,9 +1,5 @@
 package com.example.alpacamusclemaintenance.presentation.feed
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -20,20 +16,16 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.Divider
 import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.fragment.app.Fragment
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import androidx.paging.LoadState
@@ -45,37 +37,9 @@ import coil.compose.rememberImagePainter
 import com.example.alpacamusclemaintenance.domain.feed.Feed
 import com.example.alpacamusclemaintenance.presentation.R
 import com.example.alpacamusclemaintenance.presentation.WEB_VIEW_ROUTE
-import dagger.hilt.android.AndroidEntryPoint
 import org.apache.commons.lang3.time.DateFormatUtils
 import java.net.URLEncoder
 import java.nio.charset.StandardCharsets
-
-@AndroidEntryPoint
-class FeedFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View = ComposeView(requireContext()).apply {
-        setContent {
-            MaterialTheme {
-                Surface {
-//                    FeedView(
-//                        // Reference about the navigation:
-//                        //   https://developer.android.com/jetpack/compose/navigation#navigate-from-Compose
-//                        onNavigate = { feed ->
-//                            findNavController().navigate(
-//                                FeedFragmentDirections
-//                                    .actionFeedFragmentToWebViewFragment(url = feed.url)
-//                            )
-//                        }
-//                    )
-                }
-            }
-        }
-    }
-}
 
 @ExperimentalCoilApi
 @Composable
@@ -167,7 +131,9 @@ fun FeedView(
                 loadState.append is LoadState.Loading -> {
                     item {
                         Box(
-                            modifier = Modifier.fillParentMaxWidth().height(56.dp),
+                            modifier = Modifier
+                                .fillParentMaxWidth()
+                                .height(56.dp),
                             contentAlignment = Alignment.Center
                         ) {
                             CircularProgressIndicator(modifier = Modifier.size(32.dp))
