--- presentation/src/main/res/menu/navigation.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-  xmlns:android="http://schemas.android.com/apk/res/android">
-
-  <item
-    android:id="@id/homeFragment"
-    android:icon="@drawable/ic_home_black_24dp"
-    android:title="@string/title_home" />
-
-  <item
-    android:id="@id/exerciseFragment"
-    android:icon="@drawable/ic_directions_run_white_24dp"
-    android:title="@string/title_exercise" />
-
-  <item
-    android:id="@id/recordFragment"
-    android:icon="@drawable/ic_graphic_eq_white_24dp"
-    android:title="@string/title_record" />
-
-  <item
-    android:id="@id/feedFragment"
-    android:icon="@drawable/ic_library_books_white_24dp"
-    android:title="@string/title_feed" />
-
-  <item
-    android:id="@id/bugReportFragment"
-    android:icon="@drawable/ic_bug_report_white_24dp"
-    android:title="@string/title_bug_report" />
-</menu>
