--- presentation/src/main/res/menu/toolbar.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto">
-
-  <!-- "Mark Favorite", should appear as action button if possible -->
-  <item
-    android:id="@+id/action_favorite"
-    android:icon="@drawable/ic_favorite_white_24dp"
-    android:title="@string/action_favorite"
-    app:showAsAction="ifRoom" />
-
-  <!-- Settings, should always be in the overflow -->
-  <item
-    android:id="@+id/action_settings"
-    android:title="@string/action_settings"
-    app:showAsAction="never" />
-</menu>
