--- presentation/src/main/res/navigation/nav_graph.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  app:startDestination="@id/homeFragment">
-
-  <fragment
-    android:id="@+id/homeFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.home.HomeFragment"
-    android:label="@string/title_home" />
-
-  <fragment
-    android:id="@+id/exerciseFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.excercise.ExerciseFragment"
-    android:label="@string/title_exercise" />
-
-  <fragment
-    android:id="@+id/recordFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.record.RecordFragment"
-    android:label="@string/title_record" />
-
-  <fragment
-    android:id="@+id/feedFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.feed.FeedFragment"
-    android:label="@string/title_feed">
-
-    <action
-      android:id="@+id/action_feedFragment_to_webViewFragment"
-      app:destination="@id/webViewFragment"
-      app:popUpToSaveState="true" />
-  </fragment>
-
-  <fragment
-    android:id="@+id/webViewFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.WebViewFragment"
-    android:label="WebViewFragment">
-
-    <argument android:name="url" />
-  </fragment>
-
-  <fragment
-    android:id="@+id/bugReportFragment"
-    android:name="com.example.alpacamusclemaintenance.presentation.WebViewFragment"
-    android:label="WebViewFragment">
-
-    <argument
-      android:name="url"
-      android:defaultValue="https://github.com/alpaca0984/AlpacaMuscleMaintenance" />
-  </fragment>
-</navigation>
